python/requirements.txt
# dataclasses does not exist before Python3.7, needed as dependency
dataclasses;python_version<"3.7"
humanize==3.14.0
junitparser==2.7.0
lxml==4.9.1
psutil==5.9.1
PyGithub==1.55
requests==2.27.1
urllib3==1.26.11
azure-devops==6.0.0b4
dotenv-config